Splitted AUSMT into "sources"
AUSMT is now built from ausmtsrc, that is a set of scripts
that includes other scripts (à la C).

This allows better testability (the tests have been rewritten
to be self-contained and dumping a testable AUSMT on any
Linux system).

This also allows code reuse, and (hopefully) will help
fix bugs.